requirements.txt
fastapi
uvicorn
openai
httpx
python-dotenv
rapidfuzz
python-multipart
pydub